+++ CHANGELOG.md
@@ -7,6 +7,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [1.2.3] - 2021-11-27
+### Added
+* Handling for similag processes to the ProcessHandler
+* ProcessHandler is able to launch any file
+* ProcessHandler is able to launch another application and returns it runtime data
+
 ## [1.2.2] - 2021-09-17
 ### Changed
 * Removed objects out of the Validation namespace